requirements.txt
asgiref==3.9.1
Django==5.2.5
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.16.1
sqlparse==0.5.3
tzdata==2025.2
